requirements.txt
Keras==2.4.3
youtube_dl==2020.9.20
numpy==1.18.5
tensorflow==2.3.1
pandas==1.1.3
click==7.1.2
opencv_contrib_python==4.4.0.44
matplotlib==3.3.2
PyYAML==5.3.1
scikit_learn==0.23.2
